+++ tests/Header/AcceptCharsetTest.php
@@ -5,8 +5,8 @@ namespace Tests\Innmind\Http\Header;
 
 use Innmind\Http\{
     Header\AcceptCharset,
+    Header\Accept\Charset,
     Header,
-    Header\AcceptCharsetValue,
     Header\Parameter\Quality
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
@@ -15,9 +15,12 @@ class AcceptCharsetTest extends TestCase
 {
     public function testInterface()
     {
-        $h = new AcceptCharset(
-            $v = new AcceptCharsetValue('unicode-1-1', new Quality(0.8)),
-        );
+        $h = Charset::maybe('unicode-1-1', new Quality(0.8))
+            ->map(AcceptCharset::of(...))
+            ->match(
+                static fn($header) => $header,
+                static fn() => null,
+            );
 
         $this->assertInstanceOf(Header\Custom::class, $h);
         $this->assertSame('Accept-Charset: unicode-1-1;q=0.8', $h->normalize()->toString());
@@ -25,6 +28,6 @@ class AcceptCharsetTest extends TestCase
 
     public function testWithoutValues()
     {
-        $this->assertSame('Accept-Charset: ', (new AcceptCharset)->normalize()->toString());
+        $this->assertSame('Accept-Charset: ', AcceptCharset::of()->normalize()->toString());
     }
 }